+++ domain/status/build.gradle
@@ -9,6 +9,7 @@ kotlin {
 	sourceSets {
 		commonMain {
 			dependencies {
+				api(projects.domain.london)
 				api(libs.kotlin.datetime)
 				api(libs.kotlin.serialization)
 				//implementation("io.github.oshai:kotlin-logging:6.0.1")
