+++ domain/status/src/commonMain/kotlin/net/twisterrob/travel/domain/london/status/api/FeedParser.kt
@@ -1,10 +1,11 @@
 package net.twisterrob.travel.domain.london.status.api
 
+import net.twisterrob.blt.model.LineStatuses
 import net.twisterrob.travel.domain.london.status.Feed
 import net.twisterrob.travel.domain.london.status.StatusContent
 
 interface FeedParser {
 
 	@Throws(Exception::class)
-	fun parse(feed: Feed, content: StatusContent): Any
+	fun parse(feed: Feed, content: StatusContent): LineStatuses
 }
