+++ domain/status/src/commonMain/kotlin/net/twisterrob/travel/domain/london/status/api/ParsedStatusItem.kt
@@ -1,5 +1,6 @@
 package net.twisterrob.travel.domain.london.status.api
 
+import net.twisterrob.blt.model.LineStatuses
 import net.twisterrob.travel.domain.london.status.Stacktrace
 import net.twisterrob.travel.domain.london.status.StatusItem
 
@@ -9,7 +10,7 @@ sealed class ParsedStatusItem {
 
 	data class ParsedFeed(
 		override val item: StatusItem.SuccessfulStatusItem,
-		val content: Any,
+		val content: LineStatuses,
 	) : ParsedStatusItem()
 
 	data class AlreadyFailed(
