+++ web/status-history/src/main/java/net/twisterrob/travel/statushistory/controller/LineStatusHistoryController.kt
@@ -7,9 +7,6 @@ import io.micronaut.http.annotation.Get
 import io.micronaut.http.annotation.Produces
 import io.micronaut.http.annotation.QueryValue
 import io.micronaut.views.View
-import net.twisterrob.blt.io.feeds.trackernet.LineStatusFeed
-import net.twisterrob.blt.model.LineStatus
-import net.twisterrob.blt.model.LineStatuses
 import net.twisterrob.travel.domain.london.status.Feed
 import net.twisterrob.travel.domain.london.status.api.ParsedStatusItem
 import net.twisterrob.travel.domain.london.status.api.StatusHistoryRepository
@@ -47,7 +44,7 @@ private fun ParsedStatusItem.toResult(): Result {
 	val date = Date(this.item.retrievedDate.toEpochMilliseconds())
 	return when (this) {
 		is ParsedStatusItem.ParsedFeed ->
-			Result.ContentResult(date, (this.content as LineStatusFeed).toStatus())
+			Result.ContentResult(date, this.content)
 
 		is ParsedStatusItem.AlreadyFailed ->
 			Result.ErrorResult(date, Result.ErrorResult.Error(this.item.error.stacktrace))
@@ -56,21 +53,3 @@ private fun ParsedStatusItem.toResult(): Result {
 			Result.ErrorResult(date, Result.ErrorResult.Error("Error while displaying loaded XML: ${this.error.stacktrace}"))
 	}
 }
-
-private fun LineStatusFeed.toStatus(): LineStatuses =
-	LineStatuses(
-		statuses = this.lineStatuses.map { status ->
-			LineStatus(
-				line = status.line,
-				type = status.type,
-				isActive = status.isActive,
-				branchStatuses = status.branchStatuses.map { branch ->
-					LineStatus.BranchStatus(
-						fromStation = branch.fromStation,
-						toStation = branch.toStation,
-					)
-				},
-				description = status.description,
-			)
-		}
-	)
