+++ web/status-history/src/main/java/net/twisterrob/travel/statushistory/infrastructure/FeedHandlerFeedParser.kt
@@ -2,6 +2,9 @@ package net.twisterrob.travel.statushistory.infrastructure
 
 import io.micronaut.context.annotation.Bean
 import net.twisterrob.blt.io.feeds.FeedHandler
+import net.twisterrob.blt.io.feeds.trackernet.LineStatusFeed
+import net.twisterrob.blt.model.LineStatus
+import net.twisterrob.blt.model.LineStatuses
 import net.twisterrob.travel.domain.london.status.Feed
 import net.twisterrob.travel.domain.london.status.StatusContent
 import net.twisterrob.travel.domain.london.status.api.FeedParser
@@ -11,11 +14,28 @@ import net.twisterrob.blt.io.feeds.Feed as RawFeed
 class FeedHandlerFeedParser : FeedParser {
 
 	@Throws(Exception::class)
-	override fun parse(feed: Feed, content: StatusContent): Any {
+	override fun parse(feed: Feed, content: StatusContent): LineStatuses {
 		content.content.byteInputStream().use { stream ->
-			@Suppress("NEW_INFERENCE_NO_INFORMATION_FOR_PARAMETER") // No idea how.
-			val handler: FeedHandler<*> = RawFeed.valueOf(feed.name).getHandler()
-			return handler.parse(stream)
+			val handler: FeedHandler<LineStatusFeed> = RawFeed.valueOf(feed.name).getHandler()
+			return handler.parse(stream).toStatus()
 		}
 	}
 }
+
+private fun LineStatusFeed.toStatus(): LineStatuses =
+	LineStatuses(
+		statuses = this.lineStatuses.map { status ->
+			LineStatus(
+				line = status.line,
+				type = status.type,
+				isActive = status.isActive,
+				branchStatuses = status.branchStatuses.map { branch ->
+					LineStatus.BranchStatus(
+						fromStation = branch.fromStation,
+						toStation = branch.toStation,
+					)
+				},
+				description = status.description,
+			)
+		}
+	)
